Detected the state of the Caps Lock LED indicator.
Prior this update, if you coded SHIFT+A then the host interpreted
that as either uppercase "A" or lowercase "a" depending on the
current setting of the CAPS LOCK status. Now SHIFT+A will always
be interpreted by the host as uppercase "A" because the device
knows whether the CAPS LOCK LED is on or not. If CAPS LOCK is
on, then the SHIFT modifier is reset so that the "a" is interpreted
by the host as "A".